+++ harmony-fe/src/components/appbar.tsx
@@ -1,11 +1,29 @@
-import React from 'react'
-import { AppBar, Box, Toolbar, Button, Divider } from '@mui/material'
+import React, { useState } from 'react'
+import {
+  AppBar as MuiAppBar,
+  Box,
+  Toolbar,
+  Button,
+  Divider,
+} from '@mui/material'
 
-export default function HarmonyAppBar() {
+import PostModal from './post-modal'
+
+const AppBar = () => {
+  const [open, setOpen] = useState(false)
+
+  const handleOpen = () => {
+    setOpen(true)
+  }
+
+  const handleClose = () => {
+    setOpen(false)
+  }
+
   return (
     <React.Fragment>
       <Box sx={{ flexGrow: 1 }}>
-        <AppBar className="navBar" position="sticky">
+        <MuiAppBar className="navBar" position="sticky">
           <Toolbar>
             <Button href="/home">
               <Box
@@ -26,14 +44,23 @@ export default function HarmonyAppBar() {
             <Button href="/home" className="navButton" color="inherit">
               Search
             </Button>
+
             <Divider orientation="vertical" flexItem sx={{ flexGrow: 1 }} />
 
+            <Button className="navButton" color="inherit" onClick={handleOpen}>
+              New Post
+            </Button>
+            <Divider orientation="vertical" flexItem />
             <Button href="/home" className="navButton" color="inherit">
               Login
             </Button>
+            <Divider orientation="vertical" flexItem />
           </Toolbar>
-        </AppBar>
+        </MuiAppBar>
       </Box>
+      <PostModal open={open} onClose={handleClose} />
     </React.Fragment>
   )
 }
+
+export default AppBar
